Alas,
the code worked faster, but not correctly. The time was being displayed in
GMT.

Since any kind of attempt at displaying the code *properly* will lead back
to performance losses, the fix for now is just to append "GMT" to the time
so people would not be mislead by the lack of time zone modifier.