OK,
Hopefully fixed the timestamp.

Previous version was displaying the month and day in local time zone and the
rest of it in GMT, which was kind of misleading. Fixed.

Also, the default format for a slow timestamp composer is now the same as
for the fast: MM/dd HH:mm:ss.SSS