Removed. It was added just to keep the version control anyway.
The reason that this logger wrapper was never added to production code is
that the whole JDK 1.4 native logging concept is one big mistake. It was
dragged in by its feet as a result of a political struggle and never took
off. R.I.P. and serve as an example of "how not to do" for the generations
to come.